fix(mypy): reach 0 errors on src/ — fix real bugs + honest annotations
Mines the mypy error backlog (878 -> 0) for real defects, then repairs
the rest with truthful annotations rather than suppressions:

Real bugs fixed:
- model_comparison DL branch trained on the wrong extractor
  (EnhancedXgModel.extract_features returns a single dataclass;
  DLXgModel.train/predict need dl_model.extract_features(list)->matrix),
  so DL comparison silently never ran
- ShotEvent.distance_m/angle_deg were typed float=0.0 but semantically
  optional; fabrication now impossible and to_dict/from_dict roundtrip
  preserves honest-absent None
- api_v1._not_found gains NoReturn so 404 guards actually narrow
- weather TYPE_CHECKING imports, cloud/database branch typing,
  FormationEffectiveness max-key overloads, xt_model transition grid
  type, calibration reliability-curve bin types, and ~40 more

Suppression policy held: targeted ignores only where types cannot
express the truth (run_coroutine_threadsafe(loop=None) probe,
Starlette middleware variance, cv2 colormap dtype variance).

Gates: ruff check clean, ruff format clean, mypy 0/313 files,
unit suite exit 0.

+++ src/kawkab/core/role_classifier.py
@@ -92,7 +92,7 @@ def classify_player_role(
     crosses = [e for e in player_events if e.get("type") == "cross"]
 
     def_actions = len(tackles) + len(interceptions) + len(clearances)
-    total_actions = len(player_events)
+    total_actions = float(len(player_events))
 
     # pass direction bias
     forward_passes = 0
@@ -117,11 +117,13 @@ def classify_player_role(
     shot_volume = len(shots) / max(total_actions, 1)
 
     # wide vs central
-    wide_pct = sum(
-        1
-        for e in player_events
-        if e.get("start_y", PITCH_WIDTH / 2) < PITCH_WIDTH * 0.25
-        or e.get("start_y", PITCH_WIDTH / 2) > PITCH_WIDTH * 0.75
+    wide_pct = float(
+        sum(
+            1
+            for e in player_events
+            if e.get("start_y", PITCH_WIDTH / 2) < PITCH_WIDTH * 0.25
+            or e.get("start_y", PITCH_WIDTH / 2) > PITCH_WIDTH * 0.75
+        )
     )
     wide_pct /= max(total_actions, 1)
 
@@ -158,7 +160,7 @@ def classify_player_role(
     bbm_score += forward_pct * 4
     scores["box_to_box_midfielder"] = bbm_score
 
-    wm_score = wide_pct * 7
+    wm_score = float(wide_pct * 7)
     wm_score += len(crosses) / max(total_actions, 1) * 5
     wm_score += max(0, 0.5 - avg_x_rel) * 3
     scores["wide_midfielder"] = wm_score
@@ -170,19 +172,19 @@ def classify_player_role(
     am_score += (1.0 - def_actions / max(total_actions, 1)) * 3
     scores["attacking_midfielder"] = am_score
 
-    winger_score = wide_pct * 8
+    winger_score = float(wide_pct * 8)
     winger_score += len(crosses) / max(total_actions, 1) * 6
     winger_score += max(0, avg_x_rel - 0.5) * 4
     winger_score += (1.0 - def_actions / max(total_actions, 1)) * 3
     scores["winger"] = winger_score
 
-    iff_score = wide_pct * 4
+    iff_score = float(wide_pct * 4)
     iff_score += max(0, avg_x_rel - 0.6) * 6
     iff_score += shot_volume * 6
     iff_score += (1.0 - def_actions / max(total_actions, 1)) * 2
     scores["inside_forward"] = iff_score
 
-    tf_score = max(0, avg_x_rel - 0.6) * 8
+    tf_score = float(max(0, avg_x_rel - 0.6) * 8)
     tf_score += shot_volume * 7
     tf_score += backward_pct * 2
     tf_score += (1.0 - wide_pct) * 4
